e_commerce/requirements.txt
# Django==5.2.8
# celery==5.6.2
# redis==7.
# # django-celery-results
# # django-celery-beat
djangorestframework==3.16.1
djangorestframework-simplejwt
django-filter
psycopg2-binary
Django==5.2.8
celery==5.3.6
django-redis==5.4.0
redis==5.0.4
drf-yasg==1.21.15
pandas==3.0.1
numpy==2.4.3
scikit-learn==1.8.0
django-cors-headers
textblob==0.18.0
torch==2.2.2
torchvision==0.17.2
transformers==4.40.0
Pillow==10.3.0
requests==2.31.0
